requirements.txt

streamlit
pandas
requests
matplotlib
python-dotenv
pytz
timezonefinder
streamlit-searchbox
streamlit-extras
